Move NavigationTracker to a dedicated file
